Fix %rumors integration
Hotfix for a bug with %rumors integration introduced in v0.2.1.

Splits the existing call to navigate() into two, which are called either 1) when the user publishes a post without having %pals and %rumors installed or 2) when they choose not to share a post to rumors. If they do choose to share a post to rumors, the existing behaviour takes them to the %rumors app.